fix(mknod): add explicit u32 cast to MODE_RW_UGO constant
Wrap the bitwise OR expression in parentheses and cast to u32 to ensure type consistency and resolve potential compilation issues.

+++ src/uu/mknod/src/mknod.rs
@@ -15,7 +15,8 @@ use uucore::format_usage;
 use uucore::fs::makedev;
 use uucore::translate;
 
-const MODE_RW_UGO: u32 = S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH;
+const MODE_RW_UGO: u32 =
+    (S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH) as u32;
 
 mod options {
     pub const MODE: &str = "mode";
